# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Hotel Management API
  version: 1.0.0
  description: |
    API completa para la gestión integral de un sistema hotelero que incluye:
    - Autenticación y gestión de usuarios
    - Administración de habitaciones
    - Gestión de reservaciones y pagos
    - Administración de clientes y usuarios
  contact:
    name: Desarrollador de la API
    email: alfredo.sobrados.gonzalez@gmail.com
  license:
    name: License MIT
    url: https://mit-license.org/
servers:
  - url: http://localhost:8080/api
    description: Entorno de desarrollo local
...
paths:
...
  /admin/reservations:
    get:
      tags: [Reservations]
      summary: Obtener todas las reservas (admin)
      operationId: getAllReservations
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Lista de reservas obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '404':
          description: No hay reservas registradas.
        '500':
          description: Error interno del servidor.
  /admin/reservation/{id}:
    get:
      tags: [Reservations]
      summary: Obtener reserva por ID (admin)
      operationId: getReservationById
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único de la reserva.
          schema:
            type: integer
            example: 1001
      responses:
        '200':
          description: Reserva encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          description: La reserva solicitada no existe.
        '500':
          description: Error interno del servidor.
    delete:
      tags: [Reservations]
      summary: Cancelar reserva por ID (admin)
      operationId: cancelReservation
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la reserva a cancelar.
          schema:
            type: integer
            example: 1001
      responses:
        '200':
          description: La cancelación se ha hecho correctamente.
        '400':
          description: No se ha podido cancelar la reserva.
        '500':
          description: Error interno del servidor.
  /client/reservations/my:
    get:
      tags: [Reservations]
      summary: Obtener reservas asociadas al cliente autenticado
      operationId: getReservationsByClient
      security:
        - bearerAuth: [client]
      responses:
        '200':
          description: Lista de reservas del cliente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '404':
          description: No hay reservas asociadas al cliente.
        '500':
          description: Error interno del servidor.
  /client/reservation:
    post:
      tags: [Reservations]
      summary: Crear una nueva reserva (cliente)
      operationId: createReservation
      security:
        - bearerAuth: [client]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '200':
          description: Reserva creada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          description: Error al crear la reserva.
        '500':
          description: Error interno del servidor.
  /reservation/{id}:
    put:
      tags: [Reservations]
      summary: Actualizar una reserva existente
      operationId: updateReservation
      security:
        - bearerAuth: [admin, client]
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la reserva a actualizar.
          schema:
            type: integer
            example: 1001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '200':
          description: La actualización se ha hecho correctamente.
        '400':
          description: Error al actualizar la reserva.
        '500':
          description: Error interno del servidor.
  /admin/reservation/{id}/payment:
    post:
      tags: [Payments]
      summary: Crear pago para una reserva (admin)
      operationId: createPaymentOfReservation
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la reserva para la cual se crea el pago.
          schema:
            type: integer
            example: 1001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment'
      responses:
        '200':
          description: Pago creado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Error al crear el pago.
        '500':
          description: Error interno del servidor.
  /admin/reservations/{id}/payments:
    get:
      tags: [Payments]
      summary: Obtener pagos asociados a una reserva (o cliente) (admin)
      operationId: getPaymentsByClient
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la reserva (o cliente) para la cual se obtienen los pagos.
          schema:
            type: integer
            example: 1001
      responses:
        '200':
          description: Lista de pagos obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '404':
          description: No hay pagos asociados.
        '500':
          description: Error interno del servidor.

  # ==============================================
  # Payments Endpoints
  # ==============================================
  /admin/payments:
    get:
      tags: [Payments]
      summary: Obtener todos los pagos (admin)
      operationId: getAllPayments
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Lista de pagos obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '404':
          description: No hay pagos registrados en el sistema.
        '500':
          description: Error interno del servidor.
  /admin/payment/{id}:
    get:
      tags: [Payments]
      summary: Obtener pago por ID (admin)
      operationId: getPaymentById
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del pago.
          schema:
            type: integer
            example: 5001
      responses:
        '200':
          description: Pago encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          description: El pago solicitado no existe.
        '500':
          description: Error interno del servidor.
  /admin/payment/reservation/id/{reservationId}:
    get:
      tags: [Payments]
      summary: Obtener pagos por ID de reservación (admin)
      operationId: getPaymentsByReservationId
      security:
        - bearerAuth: [admin]
      parameters:
        - name: reservationId
          in: path
          required: true
          description: ID de la reservación asociada a los pagos.
          schema:
            type: integer
            example: 1001
      responses:
        '200':
          description: Lista de pagos asociados a la reservación.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '404':
          description: No hay pagos registrados asociados a la reserva.
        '500':
          description: Error interno del servidor.
  /superadmin/payment/{id}:
    put:
      tags: [Payments]
      summary: Actualizar información de un pago (superadmin)
      operationId: updatePayment
      security:
        - bearerAuth: [superadmin]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del pago a actualizar.
          schema:
            type: integer
            example: 5001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment'
      responses:
        '200':
          description: Pago actualizado correctamente.
        '400':
          description: No se ha podido actualizar el pago.
        '500':
          description: Error interno del servidor.

    delete:
      tags: [Payments]
      summary: Eliminar un pago por ID (superadmin)
      operationId: deletePayment
      security:
        - bearerAuth: [superadmin]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del pago a eliminar.
          schema:
            type: integer
            example: 5001
      responses:
        '200':
          description: Pago eliminado correctamente.
        '400':
          description: No se ha podido eliminar el pago.
        '500':
          description: Error interno del servidor.
components:
  schemas:
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
          example: 1001
        clientId:
          type: integer
          example: 123
        roomId:
          type: integer
          example: 101
        totalPrice:
          type: double
          example: 450.67
        startDate:
          type: string
          format: date
          example: "2025-03-01"
        endDate:
          type: string
          format: date
          example: "2025-03-05"
        status:
          type: string
          enum: [CONFIRMED, PENDING, CANCELED]
          example: CONFIRMED
      required:
        - clientId
        - roomId
        - totalPrice
        - startDate
        - endDate
        - status
    Payment:
      type: object
      properties:
        id:
          type: integer
          example: 1001
        amount:
          type: number
          example: 450.75
        payment_date:
          type: date
          example: "2024-01-12"
        method:
          type: string
          enum: [CARD, CASH, TRANSFER]
          example: CARD
      required:
        - id
        - amount
        - payment_date
        - method
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []